Refactor bot into modular package
