+++ zavod/config.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+STATUS_CHANNEL_ENV = "BOT_STATUS_CHANNEL_ID"
+LAST_COMMAND_CHANNEL_CONFIG_KEY = "last_command_channel_id"
+RECIPE_FEED_CHANNEL_ID = 1423404992273977364
+RESTART_LOG_CHANNEL_ID = 1423405721998987306
